# docs/api/openapi.yaml
# +JUNTOS
# Ficheiro: docs/api/openapi.yaml
# Descrição: Especificação OpenAPI 3.1 da API +JUNTOS
# Autor: (+JUNTOS team)
# Locale: pt_PT

openapi: 3.1.0
info:
  title: +JUNTOS API
  description: API para descoberta de atividades familiares em Portugal
  version: 1.0.0
  contact:
    name: Equipa +JUNTOS
    email: dev@juntos.pt
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.juntos.pt/v1
    description: Produção
  - url: https://api-stg.juntos.pt/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Desenvolvimento

security:
  - FirebaseAuth: []

paths:
  /pois:
    get:
      summary: Listar POIs próximos
      description: Encontra pontos de interesse próximos filtrados por idade das crianças
      tags: [POIs]
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
            example: 38.7223
          description: Latitude
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
            example: -9.1393
          description: Longitude
        - name: raio
          in: query
          schema:
            type: integer
            default: 5000
            minimum: 100
            maximum: 50000
          description: Raio em metros
        - name: idadeMin
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 12
          description: Idade mínima das crianças
        - name: idadeMax
          in: query
          schema:
            type: integer
            default: 12
            minimum: 0
            maximum: 12
          description: Idade máxima das crianças
        - name: limite
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Limite de resultados
      responses:
        '200':
          description: Lista de POIs encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Poi'

  /pois/search:
    get:
      summary: Pesquisar POIs
      description: Pesquisa POIs por nome ou descrição
      tags: [POIs]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            example: "parque infantil"
          description: Termo de pesquisa
        - name: limite
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Resultados da pesquisa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Poi'

  /pois/{id}:
    get:
      summary: Obter POI por ID
      description: Retorna detalhes completos de um POI
      tags: [POIs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalhes do POI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poi'
        '404':
          description: POI não encontrado

  /events:
    get:
      summary: Listar eventos
      description: Lista eventos temporários próximos
      tags: [Eventos]
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: dataInicio
          in: query
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: dataFim
          in: query
          schema:
            type: string
            format: date
            example: "2025-12-31"
      responses:
        '200':
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'

components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT do Firebase Authentication

  schemas:
    Poi:
      type: object
      required:
        - nome
        - descricao
        - categoria
        - idadeMin
        - idadeMax
        - precoMin
        - precoMax
        - latitude
        - longitude
        - morada
        - codigoPostal
        - cidade
        - distrito
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          maxLength: 255
          example: "Parque Infantil da Gulbenkian"
        descricao:
          type: string
          example: "Parque infantil com equipamentos para várias idades"
        categoria:
          type: string
          enum: [parque, museu, praia, trilho, teatro, cinema, zoo, aquario, quinta, biblioteca]
          example: "parque"
        idadeMin:
          type: integer
          minimum: 0
          maximum: 12
          example: 2
        idadeMax:
          type: integer
          minimum: 0
          maximum: 12
          example: 10
        precoMin:
          type: number
          format: decimal
          minimum: 0
          example: 0.00
        precoMax:
          type: number
          format: decimal
          minimum: 0
          example: 5.00
        latitude:
          type: number
          format: double
          example: 38.7223
        longitude:
          type: number
          format: double
          example: -9.1393
        morada:
          type: string
          example: "Av. de Berna, 45A"
        codigoPostal:
          type: string
          pattern: '^\d{4}-\d{3}$'
          example: "1067-001"
        cidade:
          type: string
          example: "Lisboa"
        distrito:
          type: string
          example: "Lisboa"
        telefone:
          type: string
          nullable: true
          example: "+351 217 823 000"
        website:
          type: string
          format: uri
          nullable: true
          example: "https://gulbenkian.pt"
        acessibilidade:
          type: boolean
          example: true
        estacionamento:
          type: boolean
          example: true
        wc:
          type: boolean
          example: true
        interior:
          type: boolean
          example: false
        exterior:
          type: boolean
          example: true
        ativo:
          type: boolean
          example: true

    Event:
      type: object
      required:
        - nome
        - descricao
        - categoria
        - dataInicio
        - dataFim
        - latitude
        - longitude
        - morada
        - cidade
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          example: "Espetáculo de Marionetas"
        descricao:
          type: string
          example: "Espetáculo interativo para crianças dos 3 aos 8 anos"
        categoria:
          type: string
          enum: [teatro, musica, workshop, feira, circo, cinema]
        idadeMin:
          type: integer
          minimum: 0
          maximum: 12
        idadeMax:
          type: integer
          minimum: 0
          maximum: 12
        preco:
          type: number
          format: decimal
          minimum: 0
        dataInicio:
          type: string
          format: date-time
        dataFim:
          type: string
          format: date-time
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        morada:
          type: string
        cidade:
          type: string
        distrito:
          type: string
        lotacaoMaxima:
          type: integer
          nullable: true
        lotacaoAtual:
          type: integer
        reservaObrigatoria:
          type: boolean
        urlReserva:
          type: string
          format: uri
          nullable: true

    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: "https://api.juntos.pt/errors/validation-error"
        title:
          type: string
          example: "Erro de validação"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "O campo 'lat' é obrigatório"
        instance:
          type: string
          format: uri
          example: "/v1/pois?lng=-9.1393"